Add sign up form to login page
