+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 // import logo from './logo.svg';
-import Login from './Login';
+import Auth from './Auth';
 import './App.css';
 
 function App() {
@@ -8,7 +8,7 @@ function App() {
   return (
     <div className="App">
     <header className="App-header">
-      <Login></Login>
+      <Auth></Auth>
     </header>
     </div>
   );
